+++ Makefile
@@ -13,8 +13,8 @@ clean:
 	@ docker rm ${worker-container}
 
 up:
-	@ docker-compose up --build -d
+	@ docker compose up --build -d
 
 down:
-	@ docker-compose down -v
+	@ docker compose down -v
