--- activities/activities.py
@@ -1,5 +1,0 @@
-from temporalio import activity
-
-@activity.defn
-async def say_hello(name: str) -> str:
-    print(f"[Activity] Hello {name}")
