+++ activities/hello_activities.py
@@ -1,0 +1,5 @@
+from temporalio import activity
+
+@activity.defn(name="say_hello")
+async def say_hello(name: str):
+    return f"Hello {name}!"
